requirements.txt
SpeechRecognition
pyttsx3
sumy
ollama
requests
beautifulsoup4
python-nmap
user_agent
pywebio
duckduckgo-search